requirements.txt
h5py==3.9.0
matplotlib==3.7.2
mpmath==1.3.0
nptyping==2.5.0
numba==0.59.1
numpy==1.24.4
python_dateutil==2.8.2
scikit_learn==1.4.2
scipy==1.13.0
tabulate==0.9.0
